style: padroniza a importação e move os DTOs para seus respectivos módulos
