+++ api/src/http/dtos/mod.rs
@@ -1,13 +1,17 @@
 pub mod auth;
 pub mod common;
+pub mod game_system;
 pub mod request;
 pub mod session;
 pub mod table;
+pub mod table_member;
 pub mod user;
 
 pub use auth::*;
 pub use common::*;
+pub use game_system::*;
 pub use request::*;
 pub use session::*;
 pub use table::*;
+pub use table_member::*;
 pub use user::*;
