+++ api/src/http/handlers/game_system.rs
@@ -1,37 +1,16 @@
-use crate::http::{dtos::CreateTableResponse, middleware::auth::auth_middleware};
-use axum::{Json, extract::State, middleware::from_fn_with_state, response::IntoResponse};
-use domain::entities::{CreateGameSystemCommand, GameSystem, GetGameSystemCommand};
+use crate::http::{
+    dtos::{
+        CreateGameSystemRequest, CreateGameSystemRespose, CreateTableResponse, GameSystemResponse,
+    },
+    middleware::auth::auth_middleware,
+};
+use axum::{Json, extract::State, middleware::from_fn_with_state};
+use domain::entities::GetGameSystemCommand;
 use infrastructure::state::AppState;
-use serde::*;
 use shared::{Error, Result};
 use std::sync::Arc;
-use utoipa::ToSchema;
 use utoipa_axum::{router::OpenApiRouter, routes};
-use uuid::Uuid;
 use validator::Validate;
-
-#[derive(Deserialize, Serialize, ToSchema, Validate)]
-pub struct CreateGameSystemRequest {
-    #[validate(length(max = 80))]
-    name: String,
-}
-
-#[derive(Deserialize, Serialize, ToSchema, Validate)]
-pub struct CreateGameSystemRespose {
-    id: Uuid,
-}
-
-impl IntoResponse for CreateTableResponse {
-    fn into_response(self) -> axum::response::Response {
-        Json(self.id).into_response()
-    }
-}
-
-impl From<CreateGameSystemRequest> for CreateGameSystemCommand {
-    fn from(value: CreateGameSystemRequest) -> Self {
-        Self { name: value.name }
-    }
-}
 
 #[utoipa::path(
     post,
@@ -59,21 +38,6 @@ async fn create_game_system(
         .id;
 
     Ok(CreateTableResponse { id })
-}
-
-#[derive(Debug, Deserialize, Serialize, ToSchema, Validate)]
-pub struct GameSystemResponse {
-    pub id: Uuid,
-    pub name: String,
-}
-
-impl From<&GameSystem> for GameSystemResponse {
-    fn from(value: &GameSystem) -> Self {
-        Self {
-            id: value.id,
-            name: value.name.clone(),
-        }
-    }
 }
 
 #[utoipa::path(
