+++ api/src/http/handlers/table.rs
@@ -344,7 +344,7 @@ pub async fn get_sessions(
             ..Default::default()
         })
         .await?
-        .iter()
+        .into_iter()
         .map(GetSessionsResponse::from)
         .collect();
