+++ api/src/http/handlers/table_members.rs
@@ -1,33 +1,13 @@
+use crate::http::dtos::TableMemberResponse;
 use axum::{
     Json,
     extract::{Path, State},
 };
-use chrono::{DateTime, Utc};
-use domain::entities::TableMember;
 use infrastructure::state::AppState;
-use serde::{Deserialize, Serialize};
 use shared::Result;
 use std::sync::Arc;
-use utoipa::ToSchema;
 use utoipa_axum::{router::OpenApiRouter, routes};
 use uuid::Uuid;
-
-#[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq, ToSchema)]
-pub struct TableMemberResponse {
-    joined_at: DateTime<Utc>,
-    user_id: Uuid,
-    table_id: Uuid,
-}
-
-impl From<TableMember> for TableMemberResponse {
-    fn from(value: TableMember) -> Self {
-        Self {
-            joined_at: value.created_at,
-            user_id: value.user_id,
-            table_id: value.table_id,
-        }
-    }
-}
 
 #[utoipa::path(
     get,
