+++ tests/utils.rs
@@ -140,14 +140,12 @@ impl TestEnvironmentBuilder {
         let table_service = TableService::new(table_repo.clone());
 
         let table_member_repo = Arc::new(PostgresTableMemberRepository::new(self.pool.clone()));
-        let table_member_service = Arc::new(TableMemberService::new(table_member_repo.clone()));
 
         let table_request_repo = Arc::new(PostgresTableRequestRepository::new(self.pool.clone()));
         let table_request_service = TableRequestService::new(
             table_request_repo.clone(),
             table_repo.clone(),
             table_member_repo.clone(),
-            table_member_service.clone(),
         );
 
         let session_repo = Arc::new(PostgresSessionRepository::new(self.pool.clone()));
